Create new iosafety.py utility file
This version has an Alias class that tries to instantiate new _alias
attributes for each instance that inherits from it. But this gets too
complicated because the iosafety class design is inherently static so
I imagine it would be difficult to "see" the attributes dynamically.
Plus, there's no good reason why you would want to dynamically change
your parameter names in a filtering appilcation.